refactor(validator): split wasm-safe manager_loop checks into common file
DIP136 and DIP137 are pure-logic checks that don't need filesystem access,
but the previous wasm stub no-oped them entirely. Extracted all pure-logic
helpers (appendManagerLoopDiags, checkManagerLoopMissingRef,
checkManagerLoopControl, checkManagerLoopSteerContextDelimiters,
checkManagerLoopUnbounded) into lint_manager_loop_common.go with no build
tag so they compile for both wasm and non-wasm. The !wasm file now contains
only lintManagerLoop + checkManagerLoopRefExists; the wasm stub calls the
common helpers without the file-existence check. Also fixed the os.Stat
call to use errors.Is(err, fs.ErrNotExist) so permission errors are not
mis-reported as missing files.
