+++ validator/lint_manager_loop.go
@@ -3,17 +3,19 @@
 package validator
 
 import (
+	"errors"
 	"fmt"
+	"io/fs"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/2389-research/dippin-lang/ir"
 )
 
 // lintManagerLoop emits DIP135 (ref missing or file does not exist), DIP136
-// (invalid control field — negative poll_interval or max_cycles), and DIP137
-// (unbounded supervision — neither stop_condition nor max_cycles set).
+// (invalid control field — negative poll_interval, max_cycles, or reserved
+// steer_context delimiter), and DIP137 (unbounded supervision — neither
+// stop_condition nor max_cycles set).
 func lintManagerLoop(w *ir.Workflow) []Diagnostic {
 	var diags []Diagnostic
 	for _, n := range w.Nodes {
@@ -22,18 +24,9 @@ func lintManagerLoop(w *ir.Workflow) []Diagnostic {
 			continue
 		}
 		diags = appendManagerLoopDiags(diags, n, cfg)
-	}
-	return diags
-}
-
-// appendManagerLoopDiags emits all DIP135/136/137 diagnostics for a node.
-func appendManagerLoopDiags(diags []Diagnostic, n *ir.Node, cfg ir.ManagerLoopConfig) []Diagnostic {
-	if d := checkManagerLoopRef(n, cfg); d != nil {
-		diags = append(diags, *d)
-	}
-	diags = append(diags, checkManagerLoopControl(n, cfg)...)
-	if d := checkManagerLoopUnbounded(n, cfg); d != nil {
-		diags = append(diags, *d)
+		if d := checkManagerLoopRefExists(n, cfg); d != nil {
+			diags = append(diags, *d)
+		}
 	}
 	return diags
 }
@@ -51,23 +44,22 @@ func resolveManagerLoopRef(n *ir.Node, cfg ir.ManagerLoopConfig) string {
 	return resolveRefPath(cfg.SubgraphRef, n.Source.File)
 }
 
-// checkManagerLoopRef emits DIP135 if subgraph_ref is empty or points to a
-// nonexistent file (when the file path can be resolved).
-func checkManagerLoopRef(n *ir.Node, cfg ir.ManagerLoopConfig) *Diagnostic {
+// checkManagerLoopRefExists emits DIP135 when subgraph_ref points to a path
+// that can be resolved but the file does not exist. Permission errors and other
+// transient IO errors are silently ignored to avoid misreporting.
+func checkManagerLoopRefExists(n *ir.Node, cfg ir.ManagerLoopConfig) *Diagnostic {
 	if cfg.SubgraphRef == "" {
-		return &Diagnostic{
-			Code:     DIP135,
-			Severity: SeverityWarning,
-			Message:  fmt.Sprintf("manager_loop %q has no subgraph_ref", n.ID),
-			Location: n.Source,
-			Help:     "set subgraph_ref to the path of a .dip file defining the child pipeline",
-		}
+		return nil
 	}
 	resolved := resolveManagerLoopRef(n, cfg)
 	if resolved == "" {
 		return nil
 	}
 	if _, err := os.Stat(resolved); err != nil {
+		if !errors.Is(err, fs.ErrNotExist) {
+			// Permission error or transient IO — don't mis-report as missing.
+			return nil
+		}
 		return &Diagnostic{
 			Code:     DIP135,
 			Severity: SeverityWarning,
@@ -77,67 +69,4 @@ func checkManagerLoopRef(n *ir.Node, cfg ir.ManagerLoopConfig) *Diagnostic {
 		}
 	}
 	return nil
-}
-
-// checkManagerLoopControl emits DIP136 for each invalid control field.
-func checkManagerLoopControl(n *ir.Node, cfg ir.ManagerLoopConfig) []Diagnostic {
-	var out []Diagnostic
-	if cfg.PollInterval < 0 {
-		out = append(out, Diagnostic{
-			Code:     DIP136,
-			Severity: SeverityWarning,
-			Message:  fmt.Sprintf("manager_loop %q poll_interval is negative (%s)", n.ID, cfg.PollInterval),
-			Location: n.Source,
-			Help:     "use a non-negative duration such as 30s; 0 means event-driven",
-		})
-	}
-	if cfg.MaxCycles < 0 {
-		out = append(out, Diagnostic{
-			Code:     DIP136,
-			Severity: SeverityWarning,
-			Message:  fmt.Sprintf("manager_loop %q max_cycles is negative (%d)", n.ID, cfg.MaxCycles),
-			Location: n.Source,
-			Help:     "use a positive integer; 0 means unbounded",
-		})
-	}
-	out = append(out, checkManagerLoopSteerContextDelimiters(n, cfg)...)
-	return out
-}
-
-// checkManagerLoopSteerContextDelimiters emits DIP136 when a steer_context key
-// or value contains ',' or '=' — those characters are delimiters in the DOT
-// flat-attr form and would corrupt round-trips through DOT export/migrate.
-func checkManagerLoopSteerContextDelimiters(n *ir.Node, cfg ir.ManagerLoopConfig) []Diagnostic {
-	var out []Diagnostic
-	for k, v := range cfg.SteerContext {
-		if strings.ContainsAny(k, ",=") || strings.ContainsAny(v, ",=") {
-			out = append(out, Diagnostic{
-				Code:     DIP136,
-				Severity: SeverityWarning,
-				Message:  fmt.Sprintf("manager_loop %q steer_context key/value contains reserved delimiter (',' or '='): %s=%s", n.ID, k, v),
-				Location: n.Source,
-				Help:     "steer_context keys and values must not contain ',' or '=' — those characters are used as delimiters in the DOT flat-attr form",
-			})
-		}
-	}
-	return out
-}
-
-// checkManagerLoopUnbounded emits DIP137 when both stop_condition and
-// max_cycles are unset — supervision can run forever.
-func checkManagerLoopUnbounded(n *ir.Node, cfg ir.ManagerLoopConfig) *Diagnostic {
-	hasStop := cfg.StopCondition != nil && cfg.StopCondition.Raw != ""
-	// Any non-zero MaxCycles (including invalid negative) indicates the user
-	// expressed bounding intent; DIP136 owns the invalid-value diagnosis.
-	hasMax := cfg.MaxCycles != 0
-	if hasStop || hasMax {
-		return nil
-	}
-	return &Diagnostic{
-		Code:     DIP137,
-		Severity: SeverityWarning,
-		Message:  fmt.Sprintf("manager_loop %q is unbounded: no stop_condition and no max_cycles", n.ID),
-		Location: n.Source,
-		Help:     "set stop_condition (e.g., stack.child.outcome = success) or max_cycles to bound supervision",
-	}
 }
